requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask==3.1.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
pandas==2.2.0
numpy==1.26.4
matplotlib==3.8.3
pillow==10.2.0
